arggllll...
[git-p4: depot-paths = "//depot/qt/main/": change = 165648]

+++ src/core/global/qglobal.h
@@ -662,12 +662,12 @@ typedef quint64 qulonglong;
 #ifdef QT_COMPAT
 typedef qint8 Q_INT8;
 typedef quint8 Q_UINT8;
-typedef qint8 Q_INT16;
-typedef quint8 Q_UINT16;
-typedef qint8 Q_INT32;
-typedef quint8 Q_UINT32;
-typedef qint8 Q_INT64;
-typedef quint8 Q_UINT64;
+typedef qint16 Q_INT16;
+typedef quint16 Q_UINT16;
+typedef qint32 Q_INT32;
+typedef quint32 Q_UINT32;
+typedef qint64 Q_INT64;
+typedef quint64 Q_UINT64;
 
 typedef qint64 Q_LLONG;
 typedef quint64 Q_ULLONG;
